Vpython
Finished Vpython… but unsure if this is exactly how we should do it. Do
you think we need to somehow incorporate the velocity? Also, I tried
working on the “being creative” part to slow down the rocket to land at
the ISS but I kind of failed. First wanted to check this part with you.